Add files via upload
Hi Elizabeth, I have updated the flexible page, so the nav bar is moving as needed. Please update what you have for the slide images and the drop list js. I am continuing to work on the footer as well as re-design a bit about the style and css. I will help you for the js when you need, please let me know as well. Thank you